[leak] Fix leaking cards which caused segfault
Additionally, a PHYSFS_init scope guard was added so that PHYSFS_deinit
gets called at the correct time.

Additionally al_uninstall_system is already called atexit, so does not
need to be called manually.

PlayerHand now cannot be copied (because we're using unique_ptr) so this
is now explicit. This avoids really cryptic error messages.