PlayerHand: cleanup and use fewer pointers and vectors
CardFactory did not need to be passed as a pointer so it is now passed
as a reference.

cards did not need to be a vector, as it will likely always have a fixed
number of elements. It is now a std::array and is iterated over by a
range-for-loop.

All PlayerHand related files now use the style.

<includes>
["this file's header"]
"includes"

All above specified groups are also sorted alphabetically. This means
that <style> includes are placed first. Followed by the header of this
file (if any), followed by any "style" includes. All groups are
separated from surrounding code by a blank line.

PlayerHand, Player1Hand and Player2Hand don't currently need
destructors, so these have been removed (they shouldn't need to deal
with destructors anyway, and alternative methods should be considered
first before a destructor is added in the future).

Constructor element initialisers have now been placed on their own line
(to make lines shorter) and used inside PlayerHand constructor (to make
the source more idiomatic).

Calls to al_draw_text have been split so lines are shorter and easier to
read.

Unnecessary included files have been pruned.